a few refactors
