+++ packages/dom-reporter/src/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
 import React from 'react';
-import { DomReporter } from './app';
+import { DomReporter } from './DomReporter';
 
 const getWasmBinary = async () => {
   const binary = await fetch("http://localhost:8080/lessThan.spec.wasm");
